QtConcurrent: fix warning due to assignment to narrower integer type
MSVC warns due to assignment of std::distance's return value (__int64)
to int when warning C4244 is enabled.

Pick-to: 6.5

+++ src/concurrent/qtconcurrentiteratekernel.h
@@ -156,7 +156,7 @@ public:
           begin(_begin),
           end(_end),
           current(_begin),
-          iterationCount(selectIteration(IteratorCategory()) ? std::distance(_begin, _end) : 0),
+          iterationCount(selectIteration(IteratorCategory()) ? static_cast<int>(std::distance(_begin, _end)) : 0),
           forIteration(selectIteration(IteratorCategory())),
           progressReportingEnabled(true)
     {
@@ -168,7 +168,7 @@ public:
           begin(_begin),
           end(_end),
           current(_begin),
-          iterationCount(selectIteration(IteratorCategory()) ? std::distance(_begin, _end) : 0),
+          iterationCount(selectIteration(IteratorCategory()) ? static_cast<int>(std::distance(_begin, _end)) : 0),
           forIteration(selectIteration(IteratorCategory())),
           progressReportingEnabled(true),
           defaultValue(U())
@@ -181,7 +181,7 @@ public:
           begin(_begin),
           end(_end),
           current(_begin),
-          iterationCount(selectIteration(IteratorCategory()) ? std::distance(_begin, _end) : 0),
+          iterationCount(selectIteration(IteratorCategory()) ? static_cast<int>(std::distance(_begin, _end)) : 0),
           forIteration(selectIteration(IteratorCategory())),
           progressReportingEnabled(true),
           defaultValue(std::forward<U>(_defaultValue))
